Apply Text specific tests to subclasses
 * Minimum PHP version bump due to use of Traits

 * The `autoload-dev` section is required for the new trait to be
 seen by the test classes.

see #40
