+++ composer.json
@@ -9,7 +9,7 @@
     ],
     "license": "MIT",
     "require": {
-        "php": ">=5.3.0"
+        "php": ">=5.4.0"
     },
     "require-dev": {
         "illuminate/support": "4.*",
@@ -21,6 +21,9 @@
         "psr-0": {
             "AdamWathan\\Form": "src/"
         }
+    },
+    "autoload-dev": {
+        "classmap": ["tests/"]
     },
     "minimum-stability": "dev"
 }
